requirements.txt
dataclasses_json==0.5.7
Pillow==9.4.0
